+++ src/modules/ideation/application/dtos/dtos.ts
@@ -1,0 +1,8 @@
+import type {
+  IdeationBodyDto,
+  IdeationRequestDto,
+} from "@/modules/ideation/infrastructure/dtos/common.dto";
+
+export interface AddIdeationUsecaseDto
+  extends Pick<IdeationRequestDto, "teamId">,
+    IdeationBodyDto {}
