+++ src/modules/ideation/application/usecases/DeleteIdeationUseCase.ts
@@ -1,8 +1,8 @@
 import { inject } from "tsyringe";
 import { TYPES } from "@/di/types";
 import { type IdeationApiRepository } from "@/modules/ideation/domain/ports/ideationApiRepository";
-import { type DeleteIdeationRequestDto } from "@/modules/ideation/domain/dtos/request.dto";
-import { type DeleteIdeationResponseDto } from "@/modules/ideation/domain/dtos/response.dto";
+import { type DeleteIdeationRequestDto } from "@/modules/ideation/infrastructure/dtos/request.dto";
+import { type DeleteIdeationResponseDto } from "@/modules/ideation/infrastructure/dtos/response.dto";
 
 export class DeleteIdeationUseCase {
   constructor(
